+++ db/migrate/20190110095634_remove_owner_from_service.rb
@@ -1,0 +1,5 @@
+class RemoveOwnerFromService < ActiveRecord::Migration[5.2]
+  def change
+    remove_reference :services, :owner
+  end
+end