+++ spec/policies/backoffice/service_policy_spec.rb
@@ -3,41 +3,95 @@
 require "rails_helper"
 
 RSpec.describe Backoffice::ServicePolicy do
-  let(:user) { create(:user, roles: [:service_owner]) }
-
-  subject { described_class }
-
-  permissions :index?, :new?, :create? do
-    it "grants access" do
-      expect(subject).to permit(user, build(:service))
+  let(:service_portfolio_manager) { create(:user, roles: [:service_portfolio_manager]) }
+  let(:service_owner) do
+    create(:user).tap do |user|
+      service = create(:service)
+      ServiceUserRelationship.create!(user: user, service: service)
     end
   end
 
+  subject { described_class }
 
-  permissions :update? do
-    it "grants access for service owner" do
-      expect(subject).to permit(user, build(:service, owner: user))
+  permissions :index? do
+    it "grants access for service portfolio manager" do
+      expect(subject).to permit(service_portfolio_manager, build(:service))
     end
 
-    it "denies access for other users" do
-      expect(subject).to_not permit(user, build(:service))
+    it "grants access for service owner" do
+      expect(subject).to permit(service_owner, build(:service))
+    end
+  end
+
+  permissions :show? do
+    it "grants access for service portfolio manager" do
+      expect(subject).to permit(service_portfolio_manager, build(:service))
+    end
+
+
+    it "grants access for owned service" do
+      expect(subject).to permit(service_owner, service_owner.owned_services.first)
+    end
+
+    it "denies access for not owned service" do
+      expect(subject).to_not permit(service_owner, build(:service))
+    end
+  end
+
+  permissions :new?, :create?, :update? do
+    it "grants access for service portfolio manager" do
+      expect(subject).to permit(service_portfolio_manager, build(:service))
+    end
+
+    it "denies access for service owner" do
+      expect(subject).to_not permit(service_owner, build(:service))
+    end
+
+    it "grants access for service portfolio manager" do
+      expect(subject).to permit(service_portfolio_manager, build(:service))
     end
   end
 
   permissions :destroy? do
-    it "grants access for service owner" do
-      expect(subject).to permit(user, build(:service, owner: user))
+    it "grants access for service portfolio manager" do
+      expect(subject).to permit(service_portfolio_manager, build(:service))
     end
 
-    it "denies access for other users" do
-      expect(subject).to_not permit(user, build(:service))
+    it "denies access for service owner" do
+      expect(subject).to_not permit(service_owner, build(:service))
     end
 
     it "denies when service has project_items attached" do
-      service = create(:service, owner: user)
+      service = create(:service)
       create(:project_item, offer: create(:offer, service: service))
 
-      expect(subject).to_not permit(user, build(:service))
+      expect(subject).to_not permit(service_portfolio_manager, service)
+    end
+  end
+
+  context "#scope" do
+    it "returns all services for service portfolio manager" do
+      create_list(:service, 2)
+
+      scope = described_class::Scope.new(service_portfolio_manager, Service.all)
+
+      expect(scope.resolve.count).to eq(2)
+    end
+
+    it "returns only owned services for service owner" do
+      _not_owned = create(:service)
+
+      scope = described_class::Scope.new(service_owner, Service.all)
+
+      expect(scope.resolve).to contain_exactly(*service_owner.owned_services)
+    end
+
+    it "returns nothing for normal user" do
+      create(:service)
+
+      scope = described_class::Scope.new(create(:user), Service.all)
+
+      expect(scope.resolve.count).to be_zero
     end
   end
 end